[Refactor] hooks/query -> hooks/cache, 사이드바 탭리스트 폰트굵기 변경, 불필요한 컴포넌트 코드여백 삭제, count useQuery 커스텀 훅 생성
