+++ app/cafe/bookmarked/page.tsx
@@ -1,36 +1,12 @@
 'use client';
 
-import { useEffect } from 'react';
-import { useQuery } from '@tanstack/react-query';
-import { useMapStore, useUserStore } from 'utils/store';
-import { countBookmarkedCafes } from 'actions/bookmarkActions';
+import { useUserStore } from 'utils/store';
 import Head from 'next/head';
+import useCountBookmarkedQuery from 'hooks/cache/useCountBookmarkedQuery';
 
 export default function BookmarkedPage() {
   const userId = useUserStore(state => state.userId);
-  const setBookmarkedCafeCount = useMapStore(
-    state => state.setBookmarkedCafeCount,
-  );
-
-  const countBookmarkedCafeQuery = useQuery({
-    queryKey: ['bookmarkedCafeCount', userId],
-    queryFn: async () => {
-      const response = await countBookmarkedCafes(userId);
-      return response || 0;
-    },
-    enabled: !!userId && userId !== 'no-user',
-    staleTime: 1000 * 60 * 3,
-    gcTime: 1000 * 60 * 5,
-  });
-
-  useEffect(() => {
-    if (countBookmarkedCafeQuery.isSuccess)
-      setBookmarkedCafeCount(countBookmarkedCafeQuery.data);
-  }, [
-    countBookmarkedCafeQuery.data,
-    countBookmarkedCafeQuery.isSuccess,
-    setBookmarkedCafeCount,
-  ]);
+  useCountBookmarkedQuery(userId);
 
   return (
     <Head>
