+++ app/cafe/collected/page.tsx
@@ -1,36 +1,13 @@
 'use client';
 
-import { useEffect } from 'react';
-import { useQuery } from '@tanstack/react-query';
-import { useMapStore, useUserStore } from 'utils/store';
-import { countCollectedCafes } from 'actions/collectActions';
+import { useUserStore } from 'utils/store';
+
 import Head from 'next/head';
+import useCountCollectedQuery from 'hooks/cache/useCountCollectedQuery';
 
 export default function CollectedPage() {
   const userId = useUserStore(state => state.userId);
-  const setCollectedCafeCount = useMapStore(
-    state => state.setCollectedCafeCount,
-  );
-
-  const countCollectedCafeQuery = useQuery({
-    queryKey: ['collectedCafeCount', userId],
-    queryFn: async () => {
-      const response = await countCollectedCafes(userId);
-      return response || 0;
-    },
-    enabled: !!userId && userId !== 'no-user',
-    staleTime: 1000 * 60 * 3,
-    gcTime: 1000 * 60 * 5,
-  });
-
-  useEffect(() => {
-    if (countCollectedCafeQuery.isSuccess)
-      setCollectedCafeCount(countCollectedCafeQuery.data);
-  }, [
-    countCollectedCafeQuery.data,
-    countCollectedCafeQuery.isSuccess,
-    setCollectedCafeCount,
-  ]);
+  useCountCollectedQuery(userId);
 
   return (
     <Head>
