+++ hooks/cache/useCollectedInfiniteQuery.ts
